[BE] refactor: 패키지 위치 재조정  #763
