+++ backend/app/src/test/java/com/onair/hearit/app/playinghistory/application/PlayingHistoryServiceTest.java
@@ -11,6 +11,7 @@ import com.onair.hearit.app.playinghistory.dto.PlayingHistoryRequest;
 import com.onair.hearit.app.playinghistory.dto.RecentlyPlayedHearitResponse;
 import com.onair.hearit.app.playinghistory.infrastructure.buffer.PlayingHistoryBuffer;
 import com.onair.hearit.app.playinghistory.infrastructure.buffer.PlayingHistoryMapBuffer;
+import com.onair.hearit.app.playinghistory.infrastructure.converter.PlayingHistoryConverter;
 import com.onair.hearit.app.userinfo.application.UserInfoService;
 import com.onair.hearit.core.config.DataSourceConfig;
 import com.onair.hearit.core.domain.Category;
@@ -48,7 +49,7 @@ import org.springframework.transaction.annotation.Transactional;
 @Transactional(propagation = Propagation.NOT_SUPPORTED)
 @Import({DbHelper.class, TestJpaAuditingConfig.class, DataSourceConfig.class, PlayingHistoryMapBuffer.class,
         PlayingHistoryCommandRepository.class, PlayingHistoryService.class, UserInfoService.class,
-        com.onair.hearit.app.playinghistory.infrastructure.buffer.converter.PlayingHistoryConverter.class})
+        PlayingHistoryConverter.class})
 class PlayingHistoryServiceTest {
 
     @Autowired
